fix: refactor CopyButton into useCopiable hook and button
