+++ src/components/CopyButton.js
@@ -1,9 +1,9 @@
-import React, { useCallback, useEffect, useRef, useState } from 'react';
-import { LinkButton } from 'indigo-react';
+import React from 'react';
 import cn from 'classnames';
-import copy from 'copy-to-clipboard';
+import { LinkButton } from 'indigo-react';
 
-const COPY_DELAY = 2 * 1000; // ms
+import useCopiable from 'lib/useCopiable';
+
 const renderDefaultCopyText = didCopy => (didCopy ? 'Copied!' : 'Copy');
 
 export default function CopyButton({
@@ -13,20 +13,7 @@ export default function CopyButton({
   className,
   ...rest
 }) {
-  const [didCopy, setDidCopy] = useState(false);
-
-  // what we really want here is a useDebouncedCallback from use-debounced
-  // but that's a relatively large dep for a single simple usage so ¯\_(ツ)_/¯
-  const listener = useRef(); // <NodeJS.Listener>
-  useEffect(() => () => clearTimeout(listener.current), []);
-
-  const doCopy = useCallback(() => {
-    copy(text);
-    setDidCopy(true);
-
-    clearTimeout(listener.current);
-    listener.current = setTimeout(() => setDidCopy(false), COPY_DELAY);
-  }, [text]);
+  const [doCopy, didCopy] = useCopiable(text);
 
   return (
     <As
